requirements.txt
flask==2.3.3
python-telegram-bot==20.6
requests==2.31.0
python-dotenv==1.0.0
werkzeug==2.3.7 